+++ bases/niagads/open_access_api/__init__.py
@@ -1,0 +1,3 @@
+from niagads.open_access_api import core
+
+__all__ = ["core"]
