--- bases/niagads/open_access_root_api/__init__.py
@@ -1,3 +1,0 @@
-from niagads.open_access_root_api import core
-
-__all__ = ["core"]
